// src/@types/Admin.d.ts
type booleanvaluesProps={
    modal:boolean,
    userid:string,
    addmodal:boolean,
    filt:boolean,
    userid:string,
    editmodal:boolean
}
export type booleanvaluableProps={
    auth:booleanvaluesProps
}
export type admindataProps={
    name:string,
    userName: string,
    phoneNumber: string,
    userId:string


}